Int to Long

+++ src/main/java/com/app_pedidos/controller/PedidoController.java
@@ -26,12 +26,12 @@ public class PedidoController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public void delete(@PathVariable Long id) {
         repository.deleteById(id);
     }
 
     @PutMapping("/{id}")
-    public void update(@PathVariable Integer id, @RequestBody Pedido pedido) {
+    public void update(@PathVariable Long id, @RequestBody Pedido pedido) {
         Pedido pedidoPesquisado = repository.getOne(id);
         if (pedidoPesquisado != null) {
             pedidoPesquisado.setData(pedido.getData());
